fix(mobile): fix rating feedback visibility and deduplicate date formatting
