+++ mobile/app/src/main/java/com/bounswe/group9/mobile/ui/discovery/MapView.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.viewinterop.AndroidView
 import coil.compose.AsyncImage
 import com.bounswe.group9.mobile.data.remote.EventListItemDto
+import com.bounswe.group9.mobile.ui.common.formatEventDate
 import org.maplibre.android.MapLibre
 import org.maplibre.android.annotations.IconFactory
 import org.maplibre.android.annotations.MarkerOptions
